requirements.txt
# Core ML and Data Processing
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0

# Computer Vision and Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Geometry Processing
shapely>=2.0.0

# OCR Engine
paddleocr>=2.6.0
imgaug>=0.4.0
pyclipper>=1.3.0
lmdb>=1.4.0

# Streamlit and Web App
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0

# Utility packages
tqdm>=4.64.0
requests>=2.28.0
pathlib2>=2.3.0